+++ src/azure-cli/azure/cli/command_modules/storage/operations/account.py
@@ -17,7 +17,9 @@ logger = get_logger(__name__)
 def create_storage_account(cmd, resource_group_name, account_name, sku=None, location=None, kind=None,
                            tags=None, custom_domain=None, encryption_services=None, access_tier=None, https_only=None,
                            enable_files_aadds=None, bypass=None, default_action=None, assign_identity=False,
-                           enable_large_file_share=None):
+                           enable_large_file_share=None, enable_files_adds=None, domain_name=None,
+                           net_bios_domain_name=None, forest_name=None, domain_guid=None, domain_sid=None,
+                           azure_storage_sid=None):
     StorageAccountCreateParameters, Kind, Sku, CustomDomain, AccessTier, Identity, Encryption, NetworkRuleSet = \
         cmd.get_models('StorageAccountCreateParameters', 'Kind', 'Sku', 'CustomDomain', 'AccessTier', 'Identity',
                        'Encryption', 'NetworkRuleSet')
@@ -35,10 +37,40 @@ def create_storage_account(cmd, resource_group_name, account_name, sku=None, loc
         params.identity = Identity()
     if https_only is not None:
         params.enable_https_traffic_only = https_only
+    AzureFilesIdentityBasedAuthentication = cmd.get_models('AzureFilesIdentityBasedAuthentication')
     if enable_files_aadds is not None:
-        AzureFilesIdentityBasedAuthentication = cmd.get_models('AzureFilesIdentityBasedAuthentication')
         params.azure_files_identity_based_authentication = AzureFilesIdentityBasedAuthentication(
             directory_service_options='AADDS' if enable_files_aadds else 'None')
+    if enable_files_adds is not None:
+        from knack.util import CLIError
+        ActiveDirectoryProperties = cmd.get_models('ActiveDirectoryProperties')
+        if enable_files_adds:  # enable AD
+            if not (domain_name and net_bios_domain_name and forest_name and domain_guid and domain_sid and
+                    azure_storage_sid):
+                raise CLIError("To enable ActiveDirectoryDomainServicesForFile, user must specify all of: "
+                               "--domain-name, --net-bios-domain-name, --forest-name, --domain-guid, --domain-sid and "
+                               "--azure_storage_sid arguments in Azure Active Directory Properties Argument group.")
+
+            active_directory_properties = ActiveDirectoryProperties(domain_name=domain_name,
+                                                                    net_bios_domain_name=net_bios_domain_name,
+                                                                    forest_name=forest_name, domain_guid=domain_guid,
+                                                                    domain_sid=domain_sid,
+                                                                    azure_storage_sid=azure_storage_sid)
+            # TODO: Enabling AD will automatically disable AADDS. Maybe we should throw error message
+
+            params.azure_files_identity_based_authentication = AzureFilesIdentityBasedAuthentication(
+                directory_service_options='AD',
+                active_directory_properties=active_directory_properties)
+
+        else:  # disable AD
+            if domain_name or net_bios_domain_name or forest_name or domain_guid or domain_sid or azure_storage_sid:  # pylint: disable=too-many-boolean-expressions
+                raise CLIError("To disable ActiveDirectoryDomainServicesForFile, user can't specify any of: "
+                               "--domain-name, --net-bios-domain-name, --forest-name, --domain-guid, --domain-sid and "
+                               "--azure_storage_sid arguments in Azure Active Directory Properties Argument group.")
+
+            params.azure_files_identity_based_authentication = AzureFilesIdentityBasedAuthentication(
+                directory_service_options='None')
+
     if enable_large_file_share:
         LargeFileSharesState = cmd.get_models('LargeFileSharesState')
         params.large_file_shares_state = LargeFileSharesState("Enabled")
@@ -109,11 +141,20 @@ def show_storage_account_usage_no_location(cmd):
     return next((x for x in scf.usage.list() if x.name.value == 'StorageAccounts'), None)  # pylint: disable=no-member
 
 
-# pylint: disable=too-many-locals
+def get_storage_account_properties(cli_ctx, account_id):
+    scf = storage_client_factory(cli_ctx)
+    from msrestazure.tools import parse_resource_id
+    result = parse_resource_id(account_id)
+    return scf.storage_accounts.get_properties(result['resource_group'], result['name'])
+
+
+# pylint: disable=too-many-locals, too-many-statements, too-many-branches, too-many-boolean-expressions
 def update_storage_account(cmd, instance, sku=None, tags=None, custom_domain=None, use_subdomain=None,
                            encryption_services=None, encryption_key_source=None, encryption_key_vault_properties=None,
                            access_tier=None, https_only=None, enable_files_aadds=None, assign_identity=False,
-                           bypass=None, default_action=None, enable_large_file_share=None):
+                           bypass=None, default_action=None, enable_large_file_share=None, enable_files_adds=None,
+                           domain_name=None, net_bios_domain_name=None, forest_name=None, domain_guid=None,
+                           domain_sid=None, azure_storage_sid=None):
     StorageAccountUpdateParameters, Sku, CustomDomain, AccessTier, Identity, Encryption, NetworkRuleSet = \
         cmd.get_models('StorageAccountUpdateParameters', 'Sku', 'CustomDomain', 'AccessTier', 'Identity',
                        'Encryption', 'NetworkRuleSet')
@@ -144,10 +185,70 @@ def update_storage_account(cmd, instance, sku=None, tags=None, custom_domain=Non
         access_tier=AccessTier(access_tier) if access_tier is not None else instance.access_tier,
         enable_https_traffic_only=https_only if https_only is not None else instance.enable_https_traffic_only
     )
+    AzureFilesIdentityBasedAuthentication = cmd.get_models('AzureFilesIdentityBasedAuthentication')
+    from knack.util import CLIError
     if enable_files_aadds is not None:
-        AzureFilesIdentityBasedAuthentication = cmd.get_models('AzureFilesIdentityBasedAuthentication')
-        params.azure_files_identity_based_authentication = AzureFilesIdentityBasedAuthentication(
-            directory_service_options='AADDS' if enable_files_aadds else 'None')
+        if enable_files_aadds:  # enable AADDS
+            origin_storage_account = get_storage_account_properties(cmd.cli_ctx, instance.id)
+            if origin_storage_account.azure_files_identity_based_authentication and \
+                    origin_storage_account.azure_files_identity_based_authentication.directory_service_options == 'AD':
+                raise CLIError("The Storage account already enabled ActiveDirectoryDomainServicesForFile, "
+                               "please disable it by running this cmdlets with \"--enable-files-adds false\" "
+                               "before enable AzureActiveDirectoryDomainServicesForFile.")
+            params.azure_files_identity_based_authentication = AzureFilesIdentityBasedAuthentication(
+                directory_service_options='AADDS' if enable_files_aadds else 'None')
+        else:  # Only disable AADDS and keep others unchanged
+            origin_storage_account = get_storage_account_properties(cmd.cli_ctx, instance.id)
+            if not origin_storage_account.azure_files_identity_based_authentication or \
+                    origin_storage_account.azure_files_identity_based_authentication.directory_service_options\
+                    == 'AADDS':
+                params.azure_files_identity_based_authentication = AzureFilesIdentityBasedAuthentication(
+                    directory_service_options='None')
+            else:
+                params.azure_files_identity_based_authentication = \
+                    origin_storage_account.azure_files_identity_based_authentication
+
+    if enable_files_adds is not None:
+        ActiveDirectoryProperties = cmd.get_models('ActiveDirectoryProperties')
+        if enable_files_adds:  # enable AD
+            if not(domain_name and net_bios_domain_name and forest_name and domain_guid and domain_sid and
+                   azure_storage_sid):
+                raise CLIError("To enable ActiveDirectoryDomainServicesForFile, user must specify all of: "
+                               "--domain-name, --net-bios-domain-name, --forest-name, --domain-guid, --domain-sid and "
+                               "--azure_storage_sid arguments in Azure Active Directory Properties Argument group.")
+            origin_storage_account = get_storage_account_properties(cmd.cli_ctx, instance.id)
+            if origin_storage_account.azure_files_identity_based_authentication and \
+                    origin_storage_account.azure_files_identity_based_authentication.directory_service_options \
+                    == 'AADDS':
+                raise CLIError("The Storage account already enabled AzureActiveDirectoryDomainServicesForFile, "
+                               "please disable it by running this cmdlets with \"--enable-files-aadds false\" "
+                               "before enable ActiveDirectoryDomainServicesForFile.")
+            active_directory_properties = ActiveDirectoryProperties(domain_name=domain_name,
+                                                                    net_bios_domain_name=net_bios_domain_name,
+                                                                    forest_name=forest_name, domain_guid=domain_guid,
+                                                                    domain_sid=domain_sid,
+                                                                    azure_storage_sid=azure_storage_sid)
+            # TODO: Enabling AD will automatically disable AADDS. Maybe we should throw error message
+
+            params.azure_files_identity_based_authentication = AzureFilesIdentityBasedAuthentication(
+                directory_service_options='AD',
+                active_directory_properties=active_directory_properties)
+
+        else:  # disable AD
+            if domain_name or net_bios_domain_name or forest_name or domain_guid or domain_sid or azure_storage_sid:
+                raise CLIError("To disable ActiveDirectoryDomainServicesForFile, user can't specify any of: "
+                               "--domain-name, --net-bios-domain-name, --forest-name, --domain-guid, --domain-sid and "
+                               "--azure_storage_sid arguments in Azure Active Directory Properties Argument group.")
+            # Only disable AD and keep others unchanged
+            origin_storage_account = get_storage_account_properties(cmd.cli_ctx, instance.id)
+            if not origin_storage_account.azure_files_identity_based_authentication or \
+                    origin_storage_account.azure_files_identity_based_authentication.directory_service_options == 'AD':
+                params.azure_files_identity_based_authentication = AzureFilesIdentityBasedAuthentication(
+                    directory_service_options='None')
+            else:
+                params.azure_files_identity_based_authentication = \
+                    origin_storage_account.azure_files_identity_based_authentication
+
     if assign_identity:
         params.identity = Identity()
     if enable_large_file_share:
@@ -164,7 +265,6 @@ def update_storage_account(cmd, instance, sku=None, tags=None, custom_domain=Non
             acl = NetworkRuleSet(bypass=bypass, virtual_network_rules=None, ip_rules=None,
                                  default_action=default_action)
         elif bypass:
-            from knack.util import CLIError
             raise CLIError('incorrect usage: --default-action ACTION [--bypass SERVICE ...]')
         params.network_rule_set = acl